requirements.txt
langchain>=0.1.0
langchain-community>=0.0.13
sentence-transformers>=2.2.2
faiss-cpu>=1.8.0
openai>=1.6.1
pypdf>=3.17.0
pdfplumber>=0.10.0
rank-bm25>=0.2.2
streamlit>=1.29.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.66.0
scikit-learn>=1.3.0

